Moves getInfos() to PwgSession class
